pairing with beth to set up home route
