+++ src/index.js
@@ -1,1 +1,5 @@
-//jhfkdjhadf
+const server = require("./server.js");
+
+const PORT = process.env.PORT || 3333;
+
+server.listen(PORT, () => console.log(`Listening on http://localhost:${PORT}`));
