+++ src/routes/sign-up.js
@@ -1,0 +1,1 @@
+const { Layout } = require('../templates/layout')
