+++ src/server.js
@@ -1,1 +1,14 @@
-//hrtakhrakjhtkg
+const express = require("express");
+const home = require("./routes/home.js");
+
+
+const server = express();
+
+server.use((req, res, next) => {
+  const time = new Date().toLocaleTimeString("en-GB");
+  console.log(`${time} ${req.method} ${req.url}`);
+  next();
+});
+
+server.get("/", home.get);
+module.exports = server;
